% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{download_mcdus_zip}
\alias{download_mcdus_zip}
\title{Download zip files.}
\usage{
download_mcdus_zip(years, save_to_wd = FALSE, timeout = 3600)
}
\arguments{
\item{years}{A character vector of years to download.}

\item{save_to_wd}{(optional) If TRUE, save to the working directory.}

\item{timeout}{(optional) Download timeout per file in seconds.}
}
\value{
TRUE if all downloads are successful, FALSE otherwise.
}
\description{
Download Mortality Multiple Cause Files in .zip format
from \insertCite{mcd}{mcdus}.
}
\references{
\insertRef{mcd}{mcdus}
}
