% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{open_mcdus_documentation}
\alias{open_mcdus_documentation}
\title{Open the data documentation from source.}
\usage{
open_mcdus_documentation(year)
}
\arguments{
\item{year}{A character string for a year (e.g., 2019).}
}
\description{
Open the data specification document from \insertCite{mcd}{mcdus}
for a specific year on a browser.
}
\references{
\insertRef{mcd}{mcdus}
}
