% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{uncompress_mcdus_zip}
\alias{uncompress_mcdus_zip}
\title{Uncompress downloaded zip files.}
\usage{
uncompress_mcdus_zip(
  years = NULL,
  save_to_wd = FALSE,
  read_from_wd = FALSE,
  unzip_method = "unzip"
)
}
\arguments{
\item{years}{(optional) A character vector of years to uncompress. Uncompress
all downloaded zip files if NULL (default).}

\item{save_to_wd}{(optional) If TRUE, uncompress to the working directory.}

\item{read_from_wd}{(optional) If TRUE, downloaded zip files are loaded from
the working directory.}

\item{unzip_method}{(optional) The method for unzip. See ?unzip.}
}
\value{
TRUE if all files are successfully uncompressed, FALSE otherwise.
}
\description{
Uncompress downloaded Mortality Multiple Cause Files in .zip format
from \insertCite{mcd}{mcdus} to a destination directory.
}
\references{
\insertRef{mcd}{mcdus}
}
